Add connections between blocks

+++ src/modules/mod.rs
@@ -3,12 +3,12 @@ pub mod builtin;
 #[derive(Default, Debug)]
 pub struct Module {
     name: String,
-    num_inputs: i32,
-    num_outputs: i32
+    num_inputs: u8,
+    num_outputs: u8
 }
 
 impl Module {
-    pub fn new(name: String, num_inputs: i32, num_outputs: i32) -> Self {
+    pub fn new(name: String, num_inputs: u8, num_outputs: u8) -> Self {
         Self {
             name,
             num_inputs,
@@ -20,11 +20,11 @@ impl Module {
         &self.name
     }
 
-    pub fn get_num_inputs(&self) -> i32 {
+    pub fn get_num_inputs(&self) -> u8 {
         self.num_inputs
     }
 
-    pub fn get_num_outputs(&self) -> i32 {
+    pub fn get_num_outputs(&self) -> u8 {
         self.num_outputs
     }
 }
